Introduced a new synapase property to skip saving artifacts to local

+++ components/mediation-admin/org.wso2.carbon.endpoint/src/main/java/org/wso2/carbon/endpoint/service/EndpointAdmin.java
@@ -12,6 +12,7 @@ import org.apache.commons.logging.LogFactory;
 import org.apache.synapse.SynapseException;
 import org.apache.synapse.config.Entry;
 import org.apache.synapse.config.SynapseConfiguration;
+import org.apache.synapse.config.SynapsePropertiesLoader;
 import org.apache.synapse.config.xml.SynapseXMLConfigurationFactory;
 import org.apache.synapse.config.xml.XMLConfigConstants;
 import org.apache.synapse.config.xml.endpoints.EndpointFactory;
@@ -69,6 +70,7 @@ public class EndpointAdmin extends AbstractServiceBusAdmin {
 
     private static final Log log = LogFactory.getLog(EndpointAdmin.class);
     public static final String WSO2_ENDPOINT_MEDIA_TYPE = "application/vnd.wso2.esb.endpoint";
+    private boolean skipLocalCopy = SynapsePropertiesLoader.getBooleanProperty("skipLocalCopy",false);
 
     /**
      * Set Endpoint status to Active
@@ -290,7 +292,7 @@ public class EndpointAdmin extends AbstractServiceBusAdmin {
             if (endpoint instanceof AbstractEndpoint) {
                 fileName = endpoint.getFileName();
             }
-            if(!Boolean.parseBoolean(System.getProperty("NonRegistryMode"))) {
+            if(!Boolean.parseBoolean(System.getProperty("NonRegistryMode")) && !skipLocalCopy) {
                 MediationPersistenceManager pm = getMediationPersistenceManager();
                 pm.deleteItem(endpointName, fileName, ServiceBusConstants.ITEM_TYPE_ENDPOINT);
             }
@@ -333,7 +335,7 @@ public class EndpointAdmin extends AbstractServiceBusAdmin {
                     if (endpoint instanceof AbstractEndpoint) {
                         fileName = endpoint.getFileName();
                     }
-                    if(!Boolean.parseBoolean(System.getProperty("NonRegistryMode"))) {
+                    if(!Boolean.parseBoolean(System.getProperty("NonRegistryMode")) && !skipLocalCopy) {
 
                         pm.deleteItem(endpointName, fileName, ServiceBusConstants.ITEM_TYPE_ENDPOINT);
                     }
@@ -378,7 +380,7 @@ public class EndpointAdmin extends AbstractServiceBusAdmin {
                         if (endpoint instanceof AbstractEndpoint) {
                             fileName = endpoint.getFileName();
                         }
-                        if(!Boolean.parseBoolean(System.getProperty("NonRegistryMode"))) {
+                        if(!Boolean.parseBoolean(System.getProperty("NonRegistryMode")) && !skipLocalCopy) {
                             MediationPersistenceManager pm = getMediationPersistenceManager();
                             pm.deleteItem(endpointName, fileName, ServiceBusConstants.ITEM_TYPE_ENDPOINT);
                         }
@@ -573,7 +575,7 @@ public class EndpointAdmin extends AbstractServiceBusAdmin {
     }
 
     private void persistEndpoint(Endpoint ep) throws EndpointAdminException {
-        if(!Boolean.parseBoolean(System.getProperty("NonRegistryMode"))) {
+        if(!Boolean.parseBoolean(System.getProperty("NonRegistryMode")) && !skipLocalCopy) {
 
             MediationPersistenceManager pm = getMediationPersistenceManager();
             pm.saveItem(ep.getName(), ServiceBusConstants.ITEM_TYPE_ENDPOINT);
